Refactor Profile component
- Move ProfileImage portion to its own component from the Profile
component.
